requirements.txt
podman-py>=3.0.0
rich>=10.0.0
pathlib>=1.0.1
python-dateutil>=2.8.2